requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
pathfinding @ git+https://github.com/LuigiLegion/python-pathfinding.git
Pillow==8.3.1
pytest==6.2.4
simplejson==3.17.3
